// src/gen/api/types/ListComments.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { Comment } from './Comment.ts'

export type ListCommentsPathParams = {
  /**
   * @description UUID identifier
   * @type string, uuid
   */
  post_id: string
}

export type ListCommentsQueryParams = {
  /**
   * @type string | undefined
   */
  cursor?: string
  /**
   * @minLength 1
   * @maxLength 50
   * @default 20
   * @type integer | undefined
   */
  limit?: number
}

/**
 * @description Cursor-paginated comments
 */
export type ListComments200 = {
  /**
   * @type array
   */
  data: Comment[]
  /**
   * @type object
   */
  pagination: {
    /**
     * @type string,null
     */
    nextCursor: string | null
    /**
     * @type boolean
     */
    hasNext: boolean
    /**
     * @type number
     */
    limit: number
  }
}

/**
 * @description Validation failed
 */
export type ListComments400 = {
  /**
   * @type string
   */
  error: string
  /**
   * @type string
   */
  message: string
  /**
   * @type string
   */
  code: string
  /**
   * @type string | undefined
   */
  requestId?: string
  /**
   * @type object | undefined
   */
  details?: {
    [key: string]: string
  }
}

/**
 * @description Missing or invalid access token
 */
export type ListComments401 = {
  /**
   * @type string
   */
  error: string
  /**
   * @type string
   */
  message: string
  /**
   * @type string
   */
  code: string
  /**
   * @type string | undefined
   */
  requestId?: string
  /**
   * @type object | undefined
   */
  details?: {
    [key: string]: string
  }
}

/**
 * @description Not found
 */
export type ListComments404 = {
  /**
   * @type string
   */
  error: string
  /**
   * @type string
   */
  message: string
  /**
   * @type string
   */
  code: string
  /**
   * @type string | undefined
   */
  requestId?: string
  /**
   * @type object | undefined
   */
  details?: {
    [key: string]: string
  }
}

/**
 * @description Internal server error
 */
export type ListComments500 = {
  /**
   * @type string
   */
  error: string
  /**
   * @type string
   */
  message: string
  /**
   * @type string
   */
  code: string
  /**
   * @type string | undefined
   */
  requestId?: string
  /**
   * @type object | undefined
   */
  details?: {
    [key: string]: string
  }
}

export type ListCommentsQueryResponse = ListComments200

export type ListCommentsQuery = {
  Response: ListComments200
  PathParams: ListCommentsPathParams
  QueryParams: ListCommentsQueryParams
  Errors: ListComments400 | ListComments401 | ListComments404 | ListComments500
}